Adiciona suporte a "lazy loading" com ILazy<> e LazyResolver<>
