+++ samples/UsingDataEntityFrameworkCore/Controllers/HomeController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using System.Diagnostics;
+using E5R.Architecture.Core;
 using E5R.Architecture.Data;
 using E5R.Architecture.Data.Abstractions;
 using E5R.Architecture.Data.Abstractions.Alias;
@@ -13,19 +14,24 @@ namespace UsingDataEntityFrameworkCore.Controllers
     {
         private readonly ILogger<HomeController> _logger;
         private readonly IStoreReader<Student> _studentStore;
+        private readonly ILazy<IStoreReader<Student>> _studentStoreLoader;
         private readonly IStoreReader<Enrollment> _enrollmentStore;
 
         public HomeController(ILogger<HomeController> logger,
                               IStoreReader<Student> studentStore,
+                              ILazy<IStoreReader<Student>> studentStoreLoader,
                               IStoreReader<Enrollment> enrollmentStore)
         {
             _logger = logger;
             _studentStore = studentStore;
+            _studentStoreLoader = studentStoreLoader;
             _enrollmentStore = enrollmentStore;
         }
 
         public IActionResult Index()
         {
+            var lazyStore = _studentStoreLoader.Value;
+            
             // Equivalentes para GetAll() com GroupBy()
             var projection = new DataProjection<Student, int, object>(
                 s => s.EnrollmentDate.Year,
@@ -37,10 +43,10 @@ namespace UsingDataEntityFrameworkCore.Controllers
                     MenorData = g.Min(c => c.EnrollmentDate)
                 });
 
-            var result = _studentStore.GetAll(projection);
+            var result = lazyStore.GetAll(projection);
             var resultList = result.ToList();
 
-            var result2 = _studentStore.AsFluentQuery()
+            var result2 = lazyStore.AsFluentQuery()
                 .Projection()
                     .GroupAndMap(s => s.EnrollmentDate.Year, g => new
                     {
@@ -54,23 +60,23 @@ namespace UsingDataEntityFrameworkCore.Controllers
             var resultList2 = result2.ToList();
 
             // Equivalentes para GetAll()
-            var a1 = _studentStore.GetAll();
-            var a2 = _studentStore.GetAll(new DataIncludes<Student>());
-            var a3 = _studentStore.AsFluentQuery().GetAll();
+            var a1 = lazyStore.GetAll();
+            var a2 = lazyStore.GetAll(new DataIncludes<Student>());
+            var a3 = lazyStore.AsFluentQuery().GetAll();
 
             // Equivalentes para GetAll() com Include() e Select()
             var a4Projection = new DataProjection<Student, object>(s => new
             {
                 StudentName = s.FirstMidName
             });
-            var a4 = _studentStore.GetAll(a4Projection);
-            var a5 = _studentStore.AsFluentQuery()
+            var a4 = lazyStore.GetAll(a4Projection);
+            var a5 = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Project()
                 .GetAll();
 
-            var a6 = _studentStore.AsFluentQuery()
+            var a6 = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Map(m => new
@@ -82,10 +88,10 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .GetAll();
 
             // Equivalentes para Find()
-            var b1 = _studentStore.Find(2, new DataIncludes<Student>());
-            var b2 = _studentStore.Find(2, null);
-            var b3 = _studentStore.Find(2);
-            var b4 = _studentStore.AsFluentQuery().Find(2);
+            var b1 = lazyStore.Find(2, new DataIncludes<Student>());
+            var b2 = lazyStore.Find(2, null);
+            var b3 = lazyStore.Find(2);
+            var b4 = lazyStore.AsFluentQuery().Find(2);
 
             // Equivalentes para Find() com Include() e Select()
             var c1 = _enrollmentStore.AsFluentQuery()
@@ -108,7 +114,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Find(2);
 
             // Equivalentes para Find() com Include() e ThenInclude()
-            var d1 = _studentStore.AsFluentQuery()
+            var d1 = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -116,7 +122,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                     .Project()
                 .Find(2);
 
-            var d2 = _studentStore.AsFluentQuery()
+            var d2 = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -135,7 +141,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Find(2);
 
             // Equivalentes para Search() com Include() e ThenInclude()
-            var e1 = _studentStore.AsFluentQuery()
+            var e1 = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -145,7 +151,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Search()
                 .ToList();
 
-            var e2 = _studentStore.AsFluentQuery()
+            var e2 = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -166,7 +172,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .ToList();
 
             // Equivalentes para Search() com Include() e ThenInclude()
-            var f1_a = _studentStore.AsFluentQuery()
+            var f1_a = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -175,7 +181,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Sort(s => s.FirstMidName)
                 .LimitedGet();
 
-            var f2_a = _studentStore.AsFluentQuery()
+            var f2_a = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -194,7 +200,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .SortDescending(s => s.FirstMidName)
                 .LimitedGet();
 
-            var f1_b = _studentStore.AsFluentQuery()
+            var f1_b = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -203,7 +209,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .OffsetBegin(3)
                 .LimitedGet();
 
-            var f2_b = _studentStore.AsFluentQuery()
+            var f2_b = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -222,7 +228,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .OffsetLimit(3)
                 .LimitedGet();
 
-            var f1_c = _studentStore.AsFluentQuery()
+            var f1_c = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -231,7 +237,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Paginate(1, 3)
                 .LimitedGet();
 
-            var f2_c = _studentStore.AsFluentQuery()
+            var f2_c = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -250,7 +256,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Paginate(2, 3)
                 .LimitedGet();
 
-            var g1_a = _studentStore.AsFluentQuery()
+            var g1_a = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -260,7 +266,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Filter(w => w.ID > 0)
                 .LimitedSearch();
 
-            var g1_b = _studentStore.AsFluentQuery()
+            var g1_b = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -270,7 +276,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Sort(s => s.LastName)
                 .LimitedSearch();
 
-            var g2_a = _studentStore.AsFluentQuery()
+            var g2_a = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
@@ -290,7 +296,7 @@ namespace UsingDataEntityFrameworkCore.Controllers
                 .Filter(w => w.ID > 0)
                 .LimitedSearch();
 
-            var g2_b = _studentStore.AsFluentQuery()
+            var g2_b = lazyStore.AsFluentQuery()
                 .Projection()
                     .Include(i => i.Enrollments)
                     .Include<Enrollment>(i => i.Enrollments)
