+++ src/E5R.Architecture.Infrastructure.AspNetCore/E5R.Architecture.Infrastructure.AspNetCore.csproj
@@ -12,8 +12,6 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.8" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.8" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Extensions\" />
