+++ src/E5R.Architecture.Infrastructure/DbTransactionUnitOfWork.cs
@@ -1,3 +1,7 @@
+// Copyright (c) E5R Development Team. All rights reserved.
+// This file is a part of E5R.Architecture.
+// Licensed under the Apache version 2.0: https://github.com/e5r/manifest/blob/master/license/APACHE-2.0.txt
+
 using System.Data.Common;
 
 namespace E5R.Architecture.Infrastructure
