+++ src/E5R.Architecture.Infrastructure/E5R.Architecture.Infrastructure.csproj
@@ -9,4 +9,8 @@
   <ItemGroup>
     <ProjectReference Include="..\E5R.Architecture.Core\E5R.Architecture.Core.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.8" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.8" />
+  </ItemGroup>
 </Project>
